+++ views/admin/modules/index.php
@@ -35,7 +35,18 @@ $this->params['breadcrumbs'][] = $this->title;
             'description',
             'class',
             'bootstrap',
-            'version',
+            [
+                'attribute' => 'version',
+                'format' => 'raw',
+                'value' => function($data) use ($module) {
+
+                    if ($new_version = $module->checkUpdates($data->name, $data->version))
+                        return $data->version . ' <label class="label label-danger">Available update to ' . $new_version . '</label>';
+                    else
+                        return $data->version;
+
+                }
+            ],
             [
                 'attribute' => 'status',
                 'format' => 'raw',